Plant-RX-main/Backend/requirements.txt
numpy==1.24.3
pandas==2.1.1
Pillow==10.0.1
flask==2.3.3
flask_cors==4.0.0
tensorflow==2.13.0
